move menu item tree kind interfaces into extension registry
